Refactor code to use packages
