+++ src/main/java/conversage/ConverSage.java
@@ -1,10 +1,10 @@
-import java.io.*;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
+package conversage;
+
+import conversage.command.Command;
+import conversage.exception.ConverSageException;
+import conversage.storage.Storage;
+import conversage.task.TaskList;
+import conversage.ui.Ui;
 
 public class ConverSage {
