+++ src/main/java/conversage/command/ByeCommand.java
@@ -1,3 +1,10 @@
+package conversage.command;
+
+import conversage.exception.ConverSageException;
+import conversage.storage.Storage;
+import conversage.task.TaskList;
+import conversage.ui.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws ConverSageException {
