+++ src/main/java/conversage/command/MarkCommand.java
@@ -1,19 +1,28 @@
-public class UnmarkCommand extends Command {
+package conversage.command;
+
+import conversage.exception.ConverSageException;
+import conversage.storage.Storage;
+import conversage.task.Task;
+import conversage.task.TaskList;
+import conversage.ui.Ui;
+
+public class MarkCommand extends Command {
     private int taskIndex;
 
     // Note that when user does "mark 1" => the 1 is a string
-    public UnmarkCommand(String taskIndex) {
+    public MarkCommand(String taskIndex) {
         this.taskIndex = Integer.parseInt(taskIndex);
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws ConverSageException {
         Task task = tasks.getTask(taskIndex - 1); // since tasks are 0-indexed
-        task.markAsUndone();
-        ui.showMessage("I've marked this task as not done yet, get to it quickly.\n" + task.toString());
+        task.markAsDone();
         ui.showLine();
-
+        ui.showMessage("Nice! I've marked this task as done:\n" + task.toString());
+        ui.showLine();
         // Update storage
         storage.save(tasks.getTasks());
     }
+    
 }
