+++ src/main/java/conversage/command/UnmarkCommand.java
@@ -1,20 +1,27 @@
-public class MarkCommand extends Command {
+package conversage.command;
+
+import conversage.exception.ConverSageException;
+import conversage.storage.Storage;
+import conversage.task.Task;
+import conversage.task.TaskList;
+import conversage.ui.Ui;
+
+public class UnmarkCommand extends Command {
     private int taskIndex;
 
     // Note that when user does "mark 1" => the 1 is a string
-    public MarkCommand(String taskIndex) {
+    public UnmarkCommand(String taskIndex) {
         this.taskIndex = Integer.parseInt(taskIndex);
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws ConverSageException {
         Task task = tasks.getTask(taskIndex - 1); // since tasks are 0-indexed
-        task.markAsDone();
+        task.markAsUndone();
+        ui.showMessage("I've marked this task as not done yet, get to it quickly.\n" + task.toString());
         ui.showLine();
-        ui.showMessage("Nice! I've marked this task as done:\n" + task.toString());
-        ui.showLine();
+
         // Update storage
         storage.save(tasks.getTasks());
     }
-    
 }
