+++ src/main/java/conversage/task/Event.java
@@ -1,4 +1,7 @@
-import java.time.LocalDate;
+package conversage.task;
+
+import conversage.exception.ConverSageException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
@@ -23,7 +26,7 @@ public class Event extends Task {
 
     @Override
     public String toFileFormat() {
-        return "Event | " + (isDone ? "Done" : "Not Done") + " | " + taskDesc + " | " + from.format(INPUT_FORMAT) + " | " + to.format(INPUT_FORMAT);
+        return "conversage.task.Event | " + (isDone ? "Done" : "Not Done") + " | " + taskDesc + " | " + from.format(INPUT_FORMAT) + " | " + to.format(INPUT_FORMAT);
     }
 
     @Override
