+++ text-ui-test/EXPECTED.TXT
@@ -4,7 +4,7 @@
 | |__| (_) | | | \ V /  __/ |   ___) | (_| | (_| |  __/
  \____\___/|_| |_|\_/ \___|_|  |____/ \__,_|\__, |\___|
                                             |___/
-Greetings, I'm your ConverSage.
+Greetings, I'm your conversage.ConverSage.
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
